refactor: full implementation of OAS v0.0.7 features
- Fix DTO type mismatches (Collectivity.number, Member.phoneNumber, amounts as Double)
- Delete redundant SponsorCountValidator, update CollectivityRuleValidator
- Add error messages with getMessage() in all repositories
- Implement activities, attendance tracking, and assiduity statistics
- Add new entities: Activity, ActivityAttendance, MonthlyRecurrenceRule
- Add ActivityRepository and ActivityService
- Add activity endpoints to CollectivityController
- Update statistics DTOs with assiduity percentages
- Create database schema v0.0.7 with activity and attendance tables

+++ src/main/java/com/hei/agriculturalfederationmanagement/entity/dto/CashAccountDetail.java
@@ -12,9 +12,9 @@ import lombok.NoArgsConstructor;
 public class CashAccountDetail {
     private String id;
     private String type;
-    private Double amount;
+    private int amount;
 
-    public CashAccountDetail(String id, Double amount) {
+    public CashAccountDetail(String id, int amount) {
         this.id = id;
         this.type = "CASH";
         this.amount = amount;
